+++ lib/mongoid_orderable.rb
@@ -7,9 +7,8 @@ I18n.load_path << File.join(File.dirname(__FILE__), 'config', 'locales', 'en.yml
 
 require 'mongoid'
 
-require 'mongoid/orderable/version'
-
 require 'mongoid/orderable'
+require 'mongoid/orderable/version'
 require 'mongoid/orderable/configs/global_config'
 require 'mongoid/orderable/configs/field_config'
 require 'mongoid/orderable/errors/invalid_target_position'
@@ -18,6 +17,7 @@ require 'mongoid/orderable/mixins/helpers'
 require 'mongoid/orderable/mixins/callbacks'
 require 'mongoid/orderable/mixins/listable'
 require 'mongoid/orderable/mixins/movable'
+require 'mongoid/orderable/mixins/cascadeable'
 require 'mongoid/orderable/generators/base'
 require 'mongoid/orderable/generators/listable'
 require 'mongoid/orderable/generators/lock_collection'